New GUI with same functionality as before.
Implementing further features should be easier now.
